+++ src/CreateServer.php
@@ -3,29 +3,37 @@
 
 namespace onservice; 
 
-define('OnServiceVersion','1.4.0');
+define('OnServiceVersion','1.5.0');
 
 class CreateServer{
 	
-	private $serverList;
+	private $serverList = [];
 	public $version = OnServiceVersion;
 
 
 	public function __construct($server = null){
-
-
 		$numargs = func_get_args();
-		$this->serverList = $numargs;	
 		foreach ($numargs as $key => $value) {
-			$serverCurrent = $value;
-
-			if( isset($serverCurrent->namespace) ){
-				$namespace = $serverCurrent->namespace;
-				$this->$namespace = $value;
-				$serverCurrent->server = $this;
-				$serverCurrent->version = $this->version;
-			}
+			$this->attachService($value);
 		}
+	}
+
+	public function attachService($driver){
+		array_push($this->serverList, $driver);		
+
+		if( !is_callable( $driver ) && isset($driver->namespace)  ){
+			$namespace = $driver->namespace;
+			$this->$namespace = $driver;
+			$driver->server = $this;				
+			$driver->version = $this->version;
+			$className = get_class($driver);
+			$className = explode('\\', $className);
+			$className = end($className);	
+
+			if(method_exists($driver, '_init'))
+			$driver->_init($this);					
+		}
+	
 	}
 	
 	public function __get($name) {
@@ -34,7 +42,7 @@ class CreateServer{
 			if( isset($value->namespace) ){
 				return $value->$name;
 			}else{
-				die('Method not exist ['.$method.']');				
+				die('Attribute not exist ['.$name.']');				
 			}
 		}
 		
@@ -43,7 +51,10 @@ class CreateServer{
 	public function __call($method,$arguments){		
 		
 		foreach ($this->serverList as $key => $value) {				
-			if( !isset($value->namespace) ){
+			if( isset($value->namespace) ){			
+				$value->server = $this;				
+				$value->version = $this->version;
+				if(method_exists($value, $method))
 				return call_user_func_array(array($value,$method), $arguments);
 			}else{
 				die('Method not exist ['.$method.']');